+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.9.0: Symfony 7 support
+
+* Upgraded support to Symfony 6
+
+- Removed ExitCode (use Symfony\Component\Console\Command constants instead)
+- Removed support for `ContainerAwareCommand`, as it's been removed from Symfony 7
+
 ## 0.8.0: Symfony 6 support
 
 * Upgraded support to Symfony 6